requirements.txt
aiodns==3.0.0
aiogram==2.23.1
aiogram-calendar==0.1.1
aiohttp==3.8.3
aiosignal==1.2.0
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
autobahn==22.7.1
Automat==20.2.0
Babel==2.9.1
cchardet==2.1.7
certifi==2022.9.24
cffi==1.15.1
channels==3.0.5
charset-normalizer==2.1.1
constantly==15.1.0
cryptography==38.0.1
daphne==3.0.2
Django==4.1.2
django-environ==0.9.0
django-nextjs==2.2.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
emoji==2.2.0
frozenlist==1.3.1
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
magic-filter==1.0.9
multidict==6.0.2
pip-autoremove==0.10.0
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycares==4.2.2
pycodestyle==2.9.1
pycparser==2.21
PyJWT==2.6.0
pyOpenSSL==22.1.0
pytz==2022.4
requests==2.28.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.3
Twisted==22.8.0
txaio==22.2.1
typing_extensions==4.4.0
ujson==5.5.0
urllib3==1.26.12
uvloop==0.17.0
websockets==10.3
yarl==1.8.1
zope.interface==5.5.0
